--- srrc/main.py
@@ -1,4 +1,0 @@
-print(yihiii)
-print(ypuh)
-
-print(1)
